fix: column data type add import

+++ packages/server/src/models/model.warranty.ts
@@ -7,7 +7,7 @@ export class Warranty {
 
   @Column({ type: "varchar" }) name: string;
   @Column({ type: "timestamp" }) effectiveDate: Date;
-  @Column({ type: "integer" }) duration: number;
+  @Column({ type: "int8" }) duration: number;
   @Column({ type: "varchar" }) durationUnit: string;
   @Column({ type: "timestamp" }) createdAt: Date;
   @Column({ type: "timestamp" }) updatedAt: Date;
